+++ services/service-gateway/src/main/resources/application-local.yml
@@ -45,6 +45,12 @@ spring:
             - Path=/file/**
           filters:
             - StripPrefix=1
+        - id: neis-service
+          uri: http://localhost:8084
+          predicates:
+            - Path=/neis/**
+          filters:
+            - StripPrefix=1
         - id: openapi-user-service
           uri: http://localhost:8081
           predicates:
