--- services/service-meal/src/main/kotlin/com/b1nd/dodamdodam/meal/application/meal/data/response/MealResponse.kt
@@ -1,11 +1,0 @@
-package com.b1nd.dodamdodam.meal.application.meal.data.response
-
-import com.b1nd.dodamdodam.meal.domain.meal.enumeration.MealType
-import java.time.LocalDate
-
-data class MealResponse(
-    val date: LocalDate,
-    val mealType: MealType,
-    val calorie: Double,
-    val menus: List<String>,
-)
