--- services/service-meal/src/main/kotlin/com/b1nd/dodamdodam/meal/domain/meal/enumeration/MealType.kt
@@ -1,8 +1,0 @@
-package com.b1nd.dodamdodam.meal.domain.meal.enumeration
-
-enum class MealType {
-    BREAKFAST,
-    LUNCH,
-    DINNER,
-    ;
-}
