--- services/service-meal/src/main/kotlin/com/b1nd/dodamdodam/meal/infrastructure/config/MealProperties.kt
@@ -1,10 +1,0 @@
-package com.b1nd.dodamdodam.meal.infrastructure.config
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "meal")
-data class MealProperties(
-    val neisApiKey: String,
-    val eduOfficeCode: String,
-    val schoolCode: String,
-)
