--- services/service-meal/src/main/kotlin/com/b1nd/dodamdodam/meal/infrastructure/config/RestTemplateConfig.kt
@@ -1,11 +1,0 @@
-package com.b1nd.dodamdodam.meal.infrastructure.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.client.RestTemplate
-
-@Configuration
-class RestTemplateConfig {
-    @Bean
-    fun restTemplate(): RestTemplate = RestTemplate()
-}
