+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/application/meal/data/response/MealResponse.kt
@@ -1,0 +1,11 @@
+package com.b1nd.dodamdodam.neis.application.meal.data.response
+
+import com.b1nd.dodamdodam.neis.domain.meal.enumeration.MealType
+import java.time.LocalDate
+
+data class MealResponse(
+    val date: LocalDate,
+    val mealType: MealType,
+    val calorie: Double,
+    val menus: List<String>,
+)
