+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/domain/meal/enumeration/MealType.kt
@@ -1,0 +1,8 @@
+package com.b1nd.dodamdodam.neis.domain.meal.enumeration
+
+enum class MealType {
+    BREAKFAST,
+    LUNCH,
+    DINNER,
+    ;
+}
