+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/domain/meal/exception/MealExceptions.kt
@@ -1,0 +1,6 @@
+package com.b1nd.dodamdodam.neis.domain.meal.exception
+
+import com.b1nd.dodamdodam.core.common.exception.BasicException
+
+class MealNotFoundException : BasicException(MealExceptionCode.MEAL_NOT_FOUND)
+class NeisApiFetchFailedException : BasicException(MealExceptionCode.NEIS_API_FAILED)
