+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/infrastructure/config/MealProperties.kt
@@ -1,0 +1,10 @@
+package com.b1nd.dodamdodam.neis.infrastructure.config
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "meal")
+data class MealProperties(
+    val neisApiKey: String,
+    val eduOfficeCode: String,
+    val schoolCode: String,
+)
