+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/infrastructure/config/RestTemplateConfig.kt
@@ -1,0 +1,21 @@
+package com.b1nd.dodamdodam.neis.infrastructure.config
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.http.HttpRequest
+import org.springframework.http.client.ClientHttpRequestExecution
+import org.springframework.http.client.ClientHttpRequestInterceptor
+import org.springframework.web.client.RestTemplate
+
+@Configuration
+class RestTemplateConfig {
+    @Bean
+    fun restTemplate(): RestTemplate = RestTemplate().apply {
+        interceptors.add(ClientHttpRequestInterceptor { request: HttpRequest, body: ByteArray, execution: ClientHttpRequestExecution ->
+            request.headers.set("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36")
+            request.headers.set("Accept", "application/json, text/plain, */*")
+            request.headers.set("Accept-Language", "ko-KR,ko;q=0.9,en-US;q=0.8,en;q=0.7")
+            execution.execute(request, body)
+        })
+    }
+}
