+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/infrastructure/exception/MealExceptionHandler.kt
@@ -1,0 +1,28 @@
+package com.b1nd.dodamdodam.neis.infrastructure.exception
+
+import com.b1nd.dodamdodam.core.common.data.Response
+import com.b1nd.dodamdodam.core.common.exception.BasicException
+import com.b1nd.dodamdodam.core.common.exception.handler.GlobalExceptionHandler
+import org.slf4j.LoggerFactory
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice
+class MealExceptionHandler : GlobalExceptionHandler {
+    private val log = LoggerFactory.getLogger(javaClass)
+
+    @ExceptionHandler(BasicException::class)
+    override fun handleDodamException(exception: BasicException): ResponseEntity<Response<Unit>> {
+        return Response.error(exception).toResponseEntity()
+    }
+
+    @ExceptionHandler(Exception::class)
+    fun handleUnexpectedException(exception: Exception): ResponseEntity<Response<Unit>> {
+        log.error("예상하지 못한 오류가 발생했어요.", exception)
+        return ResponseEntity
+            .status(HttpStatus.INTERNAL_SERVER_ERROR)
+            .body(Response(status = 500, message = "서버 내부 오류가 발생했어요."))
+    }
+}
