+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/infrastructure/security/configuration/SecurityConfig.kt
@@ -1,0 +1,30 @@
+package com.b1nd.dodamdodam.neis.infrastructure.security.configuration
+
+import com.b1nd.dodamdodam.core.security.filter.PassportFilter
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.web.SecurityFilterChain
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+
+@Configuration
+class SecurityConfig(
+    private val passportFilter: PassportFilter
+) {
+
+    @Bean
+    fun filterChain(http: HttpSecurity): SecurityFilterChain {
+        http
+            .csrf { it.disable() }
+            .addFilterBefore(passportFilter, UsernamePasswordAuthenticationFilter::class.java)
+            .authorizeHttpRequests { auth ->
+                auth.requestMatchers(
+                    "/v3/api-docs/**",
+                    "/swagger-ui/**",
+                    "/swagger-ui.html"
+                ).permitAll()
+                auth.anyRequest().authenticated()
+            }
+        return http.build()
+    }
+}
