+++ settings.gradle.kts
@@ -38,6 +38,6 @@ include(":services:service-user")
 include("services:service-inapp")
 include(":services:service-wakeup-song")
 include(":services:service-file")
-include(":services:service-meal")
+include(":services:service-neis")
 
 rootProject.name = "dodamdodam-server"
